CLN: reorg groupby to multiple modules
