+++ pandas/core/groupby/__init__.py
@@ -1,4 +1,5 @@
 # flake8: noqa
-from pandas.core.groupby.groupby import (
-    Grouper, GroupBy, SeriesGroupBy, DataFrameGroupBy
-)
+from pandas.core.groupby.groupby import GroupBy
+from pandas.core.groupby.series import SeriesGroupBy
+from pandas.core.groupby.frame import DataFrameGroupBy
+from pandas.core.groupby.ops import Grouper
