+++ pandas/core/groupby/panel.py
@@ -1,0 +1,92 @@
+from pandas import compat
+import pandas.core.common as com
+from pandas.core.groupby.groupby import NDFrameGroupBy
+from pandas.core.groupby.frame import DataFrameGroupBy
+from pandas.core.panel import Panel
+
+
+class PanelGroupBy(NDFrameGroupBy):
+
+    def aggregate(self, arg, *args, **kwargs):
+        return super(PanelGroupBy, self).aggregate(arg, *args, **kwargs)
+
+    agg = aggregate
+
+    def _iterate_slices(self):
+        if self.axis == 0:
+            # kludge
+            if self._selection is None:
+                slice_axis = self._selected_obj.items
+            else:
+                slice_axis = self._selection_list
+            slicer = lambda x: self._selected_obj[x]
+        else:
+            raise NotImplementedError("axis other than 0 is not supported")
+
+        for val in slice_axis:
+            if val in self.exclusions:
+                continue
+
+            yield val, slicer(val)
+
+    def aggregate(self, arg, *args, **kwargs):
+        """
+        Aggregate using input function or dict of {column -> function}
+
+        Parameters
+        ----------
+        arg : function or dict
+            Function to use for aggregating groups. If a function, must either
+            work when passed a Panel or when passed to Panel.apply. If
+            pass a dict, the keys must be DataFrame column names
+
+        Returns
+        -------
+        aggregated : Panel
+        """
+        if isinstance(arg, compat.string_types):
+            return getattr(self, arg)(*args, **kwargs)
+
+        return self._aggregate_generic(arg, *args, **kwargs)
+
+    def _wrap_generic_output(self, result, obj):
+        if self.axis == 0:
+            new_axes = list(obj.axes)
+            new_axes[0] = self.grouper.result_index
+        elif self.axis == 1:
+            x, y, z = obj.axes
+            new_axes = [self.grouper.result_index, z, x]
+        else:
+            x, y, z = obj.axes
+            new_axes = [self.grouper.result_index, y, x]
+
+        result = Panel._from_axes(result, new_axes)
+
+        if self.axis == 1:
+            result = result.swapaxes(0, 1).swapaxes(0, 2)
+        elif self.axis == 2:
+            result = result.swapaxes(0, 2)
+
+        return result
+
+    def _aggregate_item_by_item(self, func, *args, **kwargs):
+        obj = self._obj_with_exclusions
+        result = {}
+
+        if self.axis > 0:
+            for item in obj:
+                try:
+                    itemg = DataFrameGroupBy(obj[item],
+                                             axis=self.axis - 1,
+                                             grouper=self.grouper)
+                    result[item] = itemg.aggregate(func, *args, **kwargs)
+                except (ValueError, TypeError):
+                    raise
+            new_axes = list(obj.axes)
+            new_axes[self.axis] = self.grouper.result_index
+            return Panel._from_axes(result, new_axes)
+        else:
+            raise ValueError("axis value must be greater than 0")
+
+    def _wrap_aggregated_output(self, output, names=None):
+        raise com.AbstractMethodError(self)
