+++ pandas/tests/groupby/test_bin_groupby.py
@@ -7,6 +7,7 @@ import numpy as np
 
 from pandas.core.dtypes.common import _ensure_int64
 from pandas import Index, isna
+from pandas.core.groupby.ops import generate_bins_generic
 from pandas.util.testing import assert_almost_equal
 import pandas.util.testing as tm
 from pandas._libs import lib, groupby, reduction
@@ -54,7 +55,6 @@ class TestBinGroupers(object):
         self.bins = np.array([3, 6], dtype=np.int64)
 
     def test_generate_bins(self):
-        from pandas.core.groupby.groupby import generate_bins_generic
         values = np.array([1, 2, 3, 4, 5, 6], dtype=np.int64)
         binner = np.array([0, 3, 6, 9], dtype=np.int64)
